+++ test/EventData/AesonSpec.hs
@@ -5,7 +5,9 @@ import IntervalAlgebra
 import EventData
 import EventData.Aeson
 import EventData.Context as HC
+import EventData.Context.Domain
 import Data.Aeson
+import Data.Maybe
 import Data.Time as DT
 import Test.Hspec
 import qualified Data.ByteString.Lazy as B
@@ -30,6 +32,12 @@ testInDay = "[\"abc\", \"2020-01-01\", \"2020-01-02\", \"Diagnosis\",\
           \{\"domain\":\"Diagnosis\",\
           \ \"time\":{\"begin\":\"2020-01-01\",\"end\":\"2020-01-02\"}}]"
 
+testInDay2 :: B.ByteString
+testInDay2 = "[\"abc\", \"2020-01-01\", null, \"Diagnosis\",\
+          \[\"someThing\"],\
+          \{\"domain\":\"Diagnosis\",\
+          \ \"time\":{\"begin\":\"2020-01-01\",\"end\":null}}]"
+
 
 testInputsDay :: B.ByteString
 testInputsDay = 
@@ -42,14 +50,34 @@ testInputsDay =
       \{\"domain\":\"Diagnosis\",\
       \ \"time\":{\"begin\":\"2020-01-05\",\"end\":\"2020-01-06\"}}]"
 
-testOutInt1 = event (beginerval 1 (0 :: Int)) (HC.context $ packConcepts ["someThing"])
-testOutInt2 = event (beginerval 1 (5 :: Int)) (HC.context $ packConcepts ["someThing"])
+testInputsDay2 :: B.ByteString
+testInputsDay2 = 
+      "[\"abc\", \"2020-01-01\", null, \"Diagnosis\",\
+      \[\"someThing\"],\
+      \{\"domain\":\"Diagnosis\",\
+      \ \"time\":{\"begin\":\"2020-01-01\",\"end\":\"2020-01-02\"}}]\n\
+      \[\"abc\", \"2020-01-05\", null, \"Diagnosis\",\
+      \[\"someThing\"],\
+      \{\"domain\":\"Diagnosis\",\
+      \ \"time\":{\"begin\":\"2020-01-05\",\"end\":\"2020-01-06\"}}]"
+
+testOutInt1 = event (beginerval 1 (0 :: Int)) (HC.context ( Just $ UnimplementedDomain () ) (packConcepts ["someThing"]))
+testOutInt2 = event (beginerval 1 (5 :: Int)) (HC.context ( Just $ UnimplementedDomain () ) (packConcepts ["someThing"]))
 
 testOutDay1 = event (beginerval 1 (fromGregorian 2020 1 1))
-                     (HC.context $ packConcepts ["someThing"])
+                     (HC.context ( Just $ UnimplementedDomain () ) (packConcepts ["someThing"]))
 testOutDay2 = event (beginerval 1 (fromGregorian 2020 1 5)) 
-               (HC.context $ packConcepts [ "someThing"])
+               (HC.context ( Just $ UnimplementedDomain () ) (packConcepts [ "someThing"]))
 
+
+dmo :: Domain
+dmo = Demographics $ DemographicsFacts (DemographicsInfo BirthYear (Just "1987"))
+
+jsonDemoTest :: B.ByteString
+jsonDemoTest = "{\"domain\":\"Demographics\",\"facts\":{\"demo\":{\"field\":\"BirthYear\",\"info\":\"1987\"}}}"
+
+jsonOtherTest :: B.ByteString
+jsonOtherTest = "{\"domain\":\"Labs\",\"facts\":{\"code\":{\"code\":\"XYZ\"}}}"
 
 spec :: Spec 
 spec = do 
@@ -60,5 +88,11 @@ spec = do
 
     it "a Day event is parsed correctly" $ 
        ( decode testInDay )  `shouldBe` (Just testOutDay1)
+    it "a Day event with missing end day is parsed correctly" $ 
+       ( decode testInDay2 )  `shouldBe` (Just testOutDay1)
     it "lines of Int events are parsed correctly" $
        (parseEventDayLines testInputsDay) `shouldBe` [testOutDay1, testOutDay2]
+    it "jsonDemoTest is parsed correctly" $
+       decode jsonDemoTest  `shouldBe` Just dmo
+    it "jsonOtherTest is parsed correctly" $
+       decode jsonOtherTest  `shouldBe` Just (UnimplementedDomain ())