+++ test/EventData/ContextSpec.hs
@@ -3,13 +3,14 @@ module EventData.ContextSpec (spec) where
 
 import EventData.Context as HC
 import Test.Hspec ( shouldBe, it, Spec )
+import Data.Maybe (Maybe(Nothing))
 import Data.Set(fromList)
 
 ctxt1 :: Context
-ctxt1 = HC.context $ packConcepts  ["c1", "c2"]
+ctxt1 = HC.context Nothing (packConcepts  ["c1", "c2"])
 
 ctxt2 :: Context
-ctxt2 = HC.context $ packConcepts ["c2", "c3"]
+ctxt2 = HC.context Nothing (packConcepts ["c2", "c3"])
 
 spec :: Spec
 spec = do
@@ -25,4 +26,4 @@ spec = do
       (ctxt1 `hasConcepts` ["c3", "c4"]) `shouldBe` False
 
     it "<> combines contexts" $
-      (ctxt1 <> ctxt2) `shouldBe` HC.context (packConcepts ["c1", "c2", "c3"])
+      (ctxt1 <> ctxt2) `shouldBe` HC.context Nothing (packConcepts ["c1", "c2", "c3"])